+++ src/main/java/com/housing/app/dto/ListingSearchRequest.java
@@ -4,8 +4,8 @@ import lombok.Data;
 @Data
 public class ListingSearchRequest {
 
-    private long latitude;
-    private long longitude;
+    private double latitude;
+    private double longitude;
     private int price;
     private int area;
     private int radius;
